// src/app/models/consulta/consulta.ts
export class IConsulta {
    idPac: number;
    idDoc: number;
    idSuc: number;
    fechaCon: Date;
    estatura: number;
    peso: number;
    masaCorp: number;
    temperatura: number;
    frecResp: number;
    presArt: number;
    frecCar: number;
    grasaCorp: number;
    masaMusc: number;
    satOxigeno: number;
    motivo: string;
    diagnostico: string;
}